requirements.txt
streamlit>=1.53
langchain>=1.2
langchain-community>=0.4
langchain-classic>=1.0
langchain-anthropic>=1.3
langchain-text-splitters>=1.1
chromadb>=1.4
pypdf>=6.0
python-dotenv>=1.0
sentence-transformers>=5.0
pytest>=9.0
